Fix get_gpu_info() (#68)
* Fix get_gpu_info()

get_cpu_info() called run_cmd() with a shell pipeline which is not supported
by subprocess.check_output() by default. Instead of setting shell=True with
all the quoting fun of running a command through the shell, add an optional
filter argument to filter the lines returned by the command.

* get_gpu_info() should check from most to least specific option

Currently get_gpu_info() checks for lscpi and only if not found it will check
for nvidia-smi. Instead it should first try nvidia-smi and only if that's not
found, resort to lspci.

* Add rocm/amd support to get_gpu_info()

* Ignore test_snapshot.json from running unit tests locally

* Fix formatting